# backend/openapi/lms-openapi.yaml
openapi: 3.0.3
info:
  title: Prodeklarant LMS API
  version: 1.0.0
  description: >-
    Minimal OpenAPI 3.0 spec for LMS features (Admin + Learner). Authentication
    uses Bearer JWT tokens. Admin-only endpoints require role=ADMIN authorization
    enforced by server middleware.
servers:
  - url: /api/v1
    description: Local API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StreamTokenRequest:
      type: object
      required: [mediaId]
      properties:
        mediaId:
          type: integer
      example:
        mediaId: 10
    StreamTokenResponse:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error: { type: string }
paths:
...
  /lessons/{lessonId}/stream-token:
    post:
      summary: Request short-lived streaming token (Learner)
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamTokenRequest'
      responses:
        '200': { content: { 'application/json': { schema: { $ref: '#/components/schemas/StreamTokenResponse' } } } }
  /media/stream/{mediaId}:
    get:
      summary: Streaming proxy endpoint (supports Range)
      description: >-
        Backend verifies session token or Authorization header and streams HLS
        or MP4 bytes. This endpoint must validate token, check revocation,
        and return appropriate Range headers.
      parameters:
        - name: mediaId
          in: path
          required: true
          schema: { type: integer }
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: 'Stream (partial content supported)' }
        '401': { $ref: '#/components/schemas/Error' }
...